Isolate lightning client logic
